Make deprecated APIs private
