+++ lib/src/react_client/factory_util.dart
@@ -7,6 +7,7 @@ import 'package:react/react_client/component_factory.dart';
 import 'package:react/react_client/js_backed_map.dart';
 import 'package:react/react_client/js_interop_helpers.dart';
 import 'package:react/react_client/react_interop.dart';
+import 'package:react/src/react_client/internal_react_interop.dart';
 
 /// Converts a list of variadic children arguments to children that should be passed to ReactJS.
 ///
