+++ lib/src/react_client/private_utils.dart
@@ -8,6 +8,7 @@ import 'package:react/react.dart' show Component2;
 import 'package:react/react_client/react_interop.dart';
 import 'package:react/src/js_interop_util.dart';
 import 'package:react/src/react_client/component_registration.dart' show registerComponent2;
+import 'package:react/src/react_client/internal_react_interop.dart';
 
 /// A flag used to cache whether React is accessible.
 ///
